Update the plugin for Wireshark 4.4
Update the plugin for Wireshark 4.4.

First, blank SSID's are now returned as
`0000000000000000000000000000000000000000000000000000000000000000`,
but we want them to be displayed as an empty string. Update the Lua
script to detect this and replace this with an empty string SSID.

Second, non-blank SSID's are now wrapped in double-quotes (`"`), but we
want them displayed without the quotes. Update the Lua script to detect
this and remove the double quotes.